greatly improved the CSS for the session page
Also added notifications to let the user know when something goes wrong

organized project files to have a bit more structure
