+++ src/public/css/session/topBar.css
@@ -1,0 +1,15 @@
+.top-bar {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    padding-left: calc(var(--side-space) * 3);
+    padding-right: calc(var(--side-space) * 3);
+    padding-top: var(--padding-small);
+    padding-bottom: var(--padding-small);
+    box-shadow: 0 0 .25vw .25vw var(--light-gray);
+}
+
+.top-bar-text {
+    font-size: 1.5vw;
+    color: var(--dark-gray);
+}
