+++ src/public/css/style.css
@@ -1,10 +1,13 @@
 /* Imports */
 @import 'committees.css';
 @import 'popup.css';
+@import 'notification.css';
 @import 'session/setup.css';
 @import 'session/navbar.css';
 @import 'session/speakersList.css';
 @import 'session/rollCall.css';
+@import 'session/sidePanel.css';
+@import 'session/topBar.css';
 /* /Imports */
 
 html, body, div, span, applet, object, iframe, h1, h2, h3, h4, h5, h6, p, blockquote, pre, a, abbr, acronym, address, big, cite, code, del, dfn, em, font, img, ins, kbd, q, s, samp, small, strike, strong, sub, sup, tt, var, b, u, i, center, dl, dt, dd, ol, ul, li, fieldset, form, label, legend, caption, input, button {
