+++ src/public/js/auth/adminAuth.js
@@ -22,6 +22,8 @@ submit.addEventListener('click', async () => {
             window.location = `${redirect}`;
         }
     } else {
-        console.log('Failed');
+        const error = await response.json();
+        const notification = new Notification(error, 'red');
+        notification.show();
     }
 });
