+++ src/server/routes/get/badGatewayRouter.js
