+++ src/views/auth/adminAuth.ejs
@@ -1,8 +1,9 @@
-<%- include ('./templates/_header.ejs', { title: 'Admin Auth'} ); %>
+<%- include ('../templates/_header.ejs', { title: 'Admin Auth'} ); %>
 <body>
     <p class="header">Authorize Access</p>
     <input class="full input item" type="text" name="adminCode" id="adminCode" placeholder="Admin Code">
     <button class="full button item" id="submit">Submit</button>
+    <script src="/js/notification.js"></script>
     <script src="/js/auth/adminAuth.js"></script>
 </body>
 </html>
