+++ src/views/auth/auth.ejs
@@ -1,4 +1,4 @@
-<%- include ('./templates/_header.ejs', { title: 'Auth'} ); %>
+<%- include ('../templates/_header.ejs', { title: 'Auth'} ); %>
 <body>
     <p class="header">Authorize Access</p>
     <% if (firstName) { %>
@@ -13,6 +13,7 @@
     <% } %>
     <input class="full input item" type="text" name="accessCode" id="accessCode" placeholder="Access Code">
     <button class="full button item" id="submit">Submit</button>
+    <script src="/js/notification.js"></script>
     <script src="/js/auth/auth.js"></script>
 </body>
 </html>
