+++ src/views/committee/index.ejs
@@ -1,4 +1,4 @@
-<%- include ('./templates/_header.ejs', { title: 'Home'} ); %>
+<%- include ('../templates/_header.ejs', { title: 'Home'} ); %>
 <body>
     <a class="full button item" href="/createCommittee">Create New Committee</a>
     <input class="full input item" type="text" name="search" id="search" placeholder="&#x1F50E;&#xFE0E; Search Committees">
