+++ src/views/config.ejs
@@ -12,6 +12,7 @@
         <input class="full input item" type="text" name="adminCode" id="adminCode" placeholder="Admin Code">
     <% } %>
     <button class="full button item" id="save">Save</button>
+    <script src="/js/notification.js"></script>
     <script src="/js/config.js"></script>
 </body>
 </html>
