+++ src/views/session/setup.ejs
@@ -14,6 +14,7 @@
         </div>
     </div>
     <script>let selectedCountries = <%- JSON.stringify(committee.countries) %>;</script>
+    <script src="/js/notification.js"></script>
     <script src="/js/session/setup.js"></script>
 </body>
 </html>
