+++ src/views/templates/_header.ejs
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="/css/style.css">
-    <link rel="icon" href="/img/MUN_PRISM_1024.png">
+    <link rel="icon" href="/img/MUN_PRISM.png">
     <title>PRISM - <%= title %></title>
-    
 </head>
