+++ src/views/test.ejs
@@ -8,23 +8,6 @@
     <link rel="stylesheet" href="/css/popup.css">
 </head>
 <body>
-    <p>test text 1</p>
-    <!--
-    <div id="popup" style="display: none;">
-        <div class="popup-blackout"></div>
-        <div class="popup-container">
-            <div class="popup-menu">
-                <p class="popup-small-header">Warning!</p>
-                <p class="popup-text">Are you sure you wish to delete this committee? This is a dangerous action that cannot be undone</p>
-                <button class="popup-button popup-text popup-blue">Cancel</button>
-                <button class="popup-button popup-text popup-red">Confirm</button>
-                <input type="text" name="" id="" class="popup-input popup-text">
-            </div>
-        </div>
-    </div>
-    -->
-    <button>test button</button>
-    <p>test text 2</p>
-    <script src="/js/popup.js"></script>
+    <script src="/js/notification.js"></script>
 </body>
 </html>
