--- src/views/timer.ejs
@@ -1,11 +1,0 @@
-<%- include ('./templates/_header.ejs', { title: 'Timer'} ); %>
-<body>
-    <div id="timer" data-startTime="60">
-        <p id="time"></p>
-        <button id="timer-toggle">Start</button>
-    </div>
-
-    <script src="/socket.io/socket.io.js"></script>
-    <script src="/js/timersManager.js"></script>
-</body>
-</html>
